Add styles for header/navbar
Using scss, I've implemented the styles for the header. For the mobile
version of the site, the links are hidden in a menu until the hamburger
menu is clicked. Then, a javascript codes allows for the displaying of
the menu and changing the icon to a close icon.

Using breakpoints, once a desktop size is chosen the menu is hidden and
a traditional navbar is displayed.

This is using a mobile first development principle.